chore: move tiles to tiles directory
